Solution

+++ app/main.py
@@ -5,13 +5,10 @@ import os
 class CleanUpFile:
     def __init__(self, filename: str) -> None:
         self.filename = filename
-        self.file = None
 
     def __enter__(self) -> CleanUpFile:
-        self.file = open(self.filename, "w")
         return self
 
-    def __exit__(self, exc_type, exc_val, exc_tb) -> None:
-        if self.file:
-            self.file.close()
-        os.remove(self.filename)
+    def __exit__(self, exc_type: any, exc_val: any, exc_tb: any) -> None:
+        if os.path.exists(self.filename):
+            os.remove(self.filename)
